Added educational model, introduced bug around multiple scrollspys
The bug seems to be caused by multiple scroll spys existing and the same scroll spy
initialization being used. It seems to initialize all of them at the same time and distort
some of the top/bottom calculations. A possible solution is to pass a class to scroll
spy initialization so it only initializes those related to that specific tab